Refactor admin helpers into concern
